refactor: move the AuthenticationPrincipal to application layer
