+++ bandlog-api/src/main/java/net/effize/bandlog/application/auth/dto/AuthenticationPrincipal.java
@@ -1,4 +1,4 @@
-package net.effize.bandlog.domain.auth.model;
+package net.effize.bandlog.application.auth.dto;
 
 import net.effize.bandlog.domain.user.model.Email;
 import net.effize.bandlog.domain.user.model.SupabaseUserId;
